pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anton Fedenko <rilpackcrays@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
uvicorn = {extras = ["standart"], version = "^0.30.3"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.31"}
aiosqlite = "^0.20.0"
pydantic-settings = "^2.3.4"
black = "^24.4.2"
alembic = "^1.13.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
